+++ src/routes/Routing.tsx
@@ -11,6 +11,7 @@ import FLog from "../components/FLog"
 import UploadTemplateText from "../components/UploadTemplateText.tsx";
 import Bids from "../views/Bids"
 import Library from "../views/Library"
+import Feedback from "../views/Feedback.tsx"
 
 function Routing() {
     return (
@@ -29,6 +30,7 @@ function Routing() {
                 <Route path="/flog" element={<FLog />} />
                 <Route path="/bids" element={<Bids />} />
                 <Route path="/library" element={<Library />} />
+                <Route path="feedback" element={<Feedback />} />
 
             </Routes>
         </div>
